Rearrange and expand tests for #972
- Align test layout with code layout.
- Test .models.__getattr__().
